requirements.txt
djongo==1.3.6
djangorestframework==3.13.1
django-cors-headers==3.12.0
sqlparse==0.2.4
